requirements.txt
beautifulsoup4==4.12.0
folium==0.14.0
geocoder==1.38.1
python-dotenv==1.0.0
Requests==2.31.0
streamlit==1.23.1
streamlit_folium==0.15.0
geopy==2.3.0
